[CORE24-123] task(setup): Created db lib using knex and used it to run migrations on app start (#10)
* [CORE24-123] task(setup): Created db lib using knex and used it to run migrations on app start

* added test for db lib

* fix(server): close db connection on shutdown

* fix(ci): separated nx test command to exclude e2e projects

* task(db): use testcontainers for starting docker compose for e2e tests

* fix: formatted eslint ignored file

* fix(ci): set timeout for the CI job, nx steps may hang instead of error sometimes

+++ apps/core-api-e2e/src/support/global-teardown.ts
@@ -6,5 +6,14 @@ module.exports = async function () {
   console.log(globalThis.__TEARDOWN_MESSAGE__);
 
   // Stop the server
-  process.kill(-global.__SERVER__.pid);
+  if (global.__SERVER__) {
+    console.log('Stopping the server...');
+    process.kill(-global.__SERVER__.pid);
+  }
+
+  // Stop and remove the Docker Compose containers
+  if (global.__ENVIRONMENT__) {
+    console.log('Stopping Docker Compose environment...');
+    await global.__ENVIRONMENT__.down();
+  }
 };
